Restructured task-manager
