+++ app/Task.hs
@@ -1,0 +1,10 @@
+module Task where
+
+data Context = Context
+  { home :: FilePath
+  , tasks :: [String] }
+
+data Task = Task
+  { name :: String
+  , priorities :: [(String, String)]
+  }
